requirements.txt
# Playbook Engine Dependencies
# Version: v0.1.1
# Core dependencies for production database and AI operations

# Database connectivity (now production-ready)
psycopg2-binary>=2.9.9

# Environment and configuration
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# OpenSearch connectivity
opensearch-py>=2.5.0

# Data processing
pandas>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-mock>=3.11.0